+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Silenced `begin` supports human readable time (Format: Jan 02 2006 3:04PM MST)
 in `sensuctl` with optional timezone. Stores the field as unix epoch time.
 
+### Fixed
+- Fixed required flags in CLI `check` and `silenced` creation and added better
+error checking.
+
 ## [2.0.0-alpha.16] - 2018-02-07
 ### Added
 - Add an e2e test for proxy check requests.